+++ include/signature_digest.h
@@ -22,7 +22,7 @@ uint8_t* prepend_digestinfo(int algorithm, uint8_t* digest);
  * choice of the hash algorithm (see padding.c). Caller owns the returned
  * pointer and must Free() it.
  */
-uint8_t* SignatureDigest(const uint8_t* buf, int len, int algorithm);
+uint8_t* SignatureDigest(const uint8_t* buf, uint64_t len, int algorithm);
 
 /* Calculates the signature on a buffer [buf] of length [len] using
  * the private RSA key file from [key_file] and signature algorithm
@@ -30,6 +30,6 @@ uint8_t* SignatureDigest(const uint8_t* buf, int len, int algorithm);
  *
  * Returns the signature. Caller owns the buffer and must Free() it.
  */
-uint8_t* SignatureBuf(const uint8_t* buf, int len, const char* key_file,
+uint8_t* SignatureBuf(const uint8_t* buf, uint64_t len, const char* key_file,
                       int algorithm);
 #endif  /* VBOOT_REFERENCE_SIGNATURE_DIGEST_H_ */
